- comments and documentation - rename to CosmosClientSql and CosmosClientGraph
